Add labelmap ROS param and "is ready" service
Added a ROS parameter to store the labelmap (database) mapping object
id (stored as string) to object label name string. This wasn't working
properly wehn stored as a list on the server and indexed into, likely
because the parameter server somehow doesn't preserve the list order,
so it had to be stored as a map<string,string>.

Also added an "is ready" service advertised by the detector once the
TensorFlow setup is complete, letting other nodes know that they can
begin to use its functionality. This was mainly a problem with getting
the labelmap from the parameter server in another node before TF was
finished initializing. Perhaps an even better setup would be for the
service to return the parameter list.

Finally, changed the namespace for the output objects message to be
private (scoped within the node's name).